agents/security_analyst/requirements.txt
mcp>=1.9,<2
langsmith>=0.3,<1
httpx>=0.28.0,<1
python-dotenv>=1.0.0,<2
